From: Heiko Noordhof
It fixes the following minor issues:

(1)
There are some stale tabs in Makefile.am and src/Makefile.am.
Though this is harmless, tabs are special in Makefiles, which make emacs
complain about them saying there are "suspicious lines" in the files.

(2)
Autoconf/automake arrange for building the package in a seperate
directory. This fails with the cpufreqd (CVS as well as 2.0.0). This
also makes the "make discheck" target fail.
Example 1:
   cd sources2
   mkdir BUILD
   cd BUILD
   ../configure
   make    # Fails.
Example 2:
   cd sources2
   ./configure
   make distcheck    # Fails.

(3)
Installing as a non-privileged user in some user-writetable directory
fails, because the "install" command for  "cpufreqd.conf" tries to
change the owner of the file. This also causes "make distcheck" to fail.