Fix segfault on reading the second battery value
The problem is that check_timeout is reset after the first battery got
updated, so the values for the second battery are never read.

This results in status->value == NULL, and strncmp segfaults.

The solution is to update check_timeout after all battery values are updated.
Patch from: Stefan Bühler

Signed-off-by: Mattia Dongili <malattia@linux.it>